refactor(protocol): extract UpdateCodec from BgpSession — Phase 1 of #93
9 internal-static codec/validator methods + 3 community-set comparers moved
from BgpSession (Server) to a new public static UpdateCodec (Protocol):
the outbound attribute builder (BuildAsPathAttributes, BuildUpdateAttributes,
CreateUpdateAttributeCache, GetCachedUpdateAttributes, WithLargeCommunity
Attribute), the inbound RFC 6793 validators/merger (ValidateMandatoryAttributes,
MergeAsPathWithAs4Path, ValidateAggregatorReconstruction), and the malformed-
capability TLV builder (GetMalformedFourOctetAsnCapabilityData). All were pure
statics — zero instance-field access, every input a parameter — and lived in
BgpSession solely as internal statics reachable via InternalsVisibleTo.

Moved to BGPLite.Protocol as public alongside AttributeHelper/BgpMessageWriter
(their natural siblings). This removes the InternalsVisibleTo test-backdoor for
the codec methods and keeps BgpSession focused on FSM/transport/timer concerns.
GroupByCommunitySet + its SameCommunitySet/PartitionByCommunitySet helpers
remain in BgpSession (Phase 2 — they are route-assembly, not codec), referencing
the now-public CommunitySetComparer/LargeCommunitySetComparer/CommunitySetPair
Comparer from Protocol.

24 unit tests repointed (BgpSession.X → UpdateCodec.X). All 14 socket tests
stay green unchanged.
