+++ BGPLite.Server/BgpSession.cs
@@ -565,9 +565,9 @@ public sealed class BgpSession : IDisposable
                     }
                 }
 
-                ValidateMandatoryAttributes(originSeen, asPathSeen, nextHopSeen);
-                asPath = MergeAsPathWithAs4Path(asPath, as4Path);
-                ValidateAggregatorReconstruction(aggregatorAsn, as4AggregatorAsn);
+                UpdateCodec.ValidateMandatoryAttributes(originSeen, asPathSeen, nextHopSeen);
+                asPath = UpdateCodec.MergeAsPathWithAs4Path(asPath, as4Path);
+                UpdateCodec.ValidateAggregatorReconstruction(aggregatorAsn, as4AggregatorAsn);
 
                 foreach (var nlri in update.Nlri)
                 {
@@ -854,7 +854,7 @@ public sealed class BgpSession : IDisposable
         // build them once per community set and reuse instead of rebuilding on each batch (#87).
         // Scoped to this send only: the cache dies with the dictionary, so it can never serve a
         // later send that carries a different nextHop or renegotiated ASN.
-        var attrCache = CreateUpdateAttributeCache();
+        var attrCache = UpdateCodec.CreateUpdateAttributeCache();
 
         foreach (var route in routes)
         {
@@ -885,159 +885,16 @@ public sealed class BgpSession : IDisposable
         // communities on the wire.
         foreach (var groupRoutes in GroupByCommunitySet(routes))
         {
-            var attrs = GetCachedUpdateAttributes(_bgpConfig.Asn, _localFourByteAsn, nextHop, groupRoutes[0].Communities, attrCache);
+            var attrs = UpdateCodec.GetCachedUpdateAttributes(_bgpConfig.Asn, _localFourByteAsn, nextHop, groupRoutes[0].Communities, attrCache);
             // LARGE_COMMUNITY is appended per group AFTER fetching the cached base attrs, so the
             // #87 cache stays keyed by regular communities only and is never mutated. Routes that
             // share regular communities but differ in large communities reuse the same base attrs
             // and only diverge in this final attribute.
-            attrs = WithLargeCommunityAttribute(attrs, groupRoutes[0].LargeCommunities);
+            attrs = UpdateCodec.WithLargeCommunityAttribute(attrs, groupRoutes[0].LargeCommunities);
 
             var nlri = groupRoutes.Select(r => new IpPrefix(r.Prefix, r.PrefixLength)).ToList();
             await SendUpdateBatchAsync(attrs, nlri);
         }
-    }
-
-    /// <summary>
-    /// Builds AS_PATH and optionally AS4_PATH attributes per RFC 6793 §6.
-    /// - <paramref name="localFourByteAsn"/>=true: single 4-byte AS_PATH.
-    /// - <paramref name="localFourByteAsn"/>=false: 2-byte AS_PATH (AS_TRANS(23456) if
-    ///   <paramref name="localAsn"/> &gt; 65535) plus AS4_PATH carrying the true 4-byte ASN.
-    /// Internal for test coverage.
-    /// </summary>
-    internal static List<PathAttribute> BuildAsPathAttributes(uint localAsn, bool localFourByteAsn)
-    {
-        var attrs = new List<PathAttribute>(2);
-        if (localFourByteAsn)
-        {
-            attrs.Add(AttributeHelper.WriteAsPath([localAsn], fourByteAsn: true));
-        }
-        else
-        {
-            var asPathAsn = localAsn > ushort.MaxValue ? BgpConstants.AsPath.AsTrans : localAsn;
-            attrs.Add(AttributeHelper.WriteAsPath([asPathAsn], fourByteAsn: false));
-
-            if (localAsn > ushort.MaxValue)
-                attrs.Add(AttributeHelper.WriteAs4Path([localAsn]));
-        }
-        return attrs;
-    }
-
-    /// <summary>
-    /// Builds outbound UPDATE path attributes in RFC order: ORIGIN, AS_PATH, NEXT_HOP,
-    /// COMMUNITY, AS4_PATH. Internal for test coverage.
-    /// </summary>
-    internal static List<PathAttribute> BuildUpdateAttributes(uint localAsn, bool localFourByteAsn, uint nextHop, uint[] communities)
-    {
-        var attrs = new List<PathAttribute>(5)
-        {
-            AttributeHelper.WriteOrigin(BgpOrigin.Igp),
-        };
-
-        var asPathAttrs = BuildAsPathAttributes(localAsn, localFourByteAsn);
-        attrs.Add(asPathAttrs[0]);
-        attrs.Add(AttributeHelper.WriteNextHop(nextHop));
-
-        if (communities.Length > 0)
-            attrs.Add(AttributeHelper.WriteCommunities(communities));
-
-        if (asPathAttrs.Count > 1)
-            attrs.Add(asPathAttrs[1]);
-
-        return attrs;
-    }
-
-    /// <summary>
-    /// Creates a per-send cache of built UPDATE path attributes, keyed by community set. The
-    /// cache is scoped to a single <see cref="SendRoutesAsync"/> invocation: the ASN/nextHop
-    /// inputs are constant for that whole send, so identical community sets yield byte-identical
-    /// <see cref="PathAttribute"/> lists that can be reused across the N 100-NLRI batches (#87).
-    /// Internal for test coverage.
-    /// </summary>
-    internal static Dictionary<uint[], List<PathAttribute>> CreateUpdateAttributeCache() =>
-        new(CommunitySetComparer.Instance);
-
-    /// <summary>
-    /// Returns the UPDATE path attributes for <paramref name="communities"/>, building them on
-    /// first request for a community set and returning the cached list thereafter. The cached
-    /// <see cref="PathAttribute"/> payloads are immutable, so the same list is safely shared by
-    /// every UPDATE emitted for that community set. Internal for test coverage.
-    /// </summary>
-    internal static List<PathAttribute> GetCachedUpdateAttributes(
-        uint localAsn, bool localFourByteAsn, uint nextHop, uint[] communities,
-        Dictionary<uint[], List<PathAttribute>> cache)
-    {
-        if (cache.TryGetValue(communities, out var cached))
-            return cached;
-
-        var attrs = BuildUpdateAttributes(localAsn, localFourByteAsn, nextHop, communities);
-        cache[communities] = attrs;
-        return attrs;
-    }
-
-    /// <summary>
-    /// Returns the path attributes for an UPDATE carrying the given Large Community set: the
-    /// cached base attributes (ORIGIN/AS_PATH/NEXT_HOP/COMMUNITY/AS4_PATH) untouched when
-    /// <paramref name="largeCommunities"/> is empty, otherwise a shallow copy with a
-    /// LARGE_COMMUNITY attribute appended. The cached base list is never mutated, so other
-    /// batches in the same send that share regular communities but carry a different (or empty)
-    /// large-community set still observe the correct base. Appended last, which keeps the
-    /// emitted attributes in ascending type-code order (32 sorts after AS4_PATH 17). Internal
-    /// for test coverage.
-    /// </summary>
-    internal static List<PathAttribute> WithLargeCommunityAttribute(
-        List<PathAttribute> baseAttrs, (uint Global, uint Local1, uint Local2)[] largeCommunities)
-    {
-        if (largeCommunities.Length == 0)
-            return baseAttrs;
-
-        var withLarge = new List<PathAttribute>(baseAttrs.Count + 1);
-        withLarge.AddRange(baseAttrs);
-        withLarge.Add(AttributeHelper.WriteLargeCommunities(largeCommunities));
-        return withLarge;
-    }
-
-    /// <summary>
-    /// Validates that a route announcement carried the mandatory well-known attributes.
-    /// Internal for test coverage.
-    /// </summary>
-    internal static void ValidateMandatoryAttributes(bool originSeen, bool asPathSeen, bool nextHopSeen)
-    {
-        if (!originSeen)
-            throw new BgpNotificationException(BgpConstants.Error.UpdateMessageError, BgpConstants.SubError.MissingWellKnownAttribute, "Missing mandatory ORIGIN attribute");
-        if (!asPathSeen)
-            throw new BgpNotificationException(BgpConstants.Error.UpdateMessageError, BgpConstants.SubError.MissingWellKnownAttribute, "Missing mandatory AS_PATH attribute");
-        if (!nextHopSeen)
-            throw new BgpNotificationException(BgpConstants.Error.UpdateMessageError, BgpConstants.SubError.MissingWellKnownAttribute, "Missing mandatory NEXT_HOP attribute");
-    }
-
-    /// <summary>
-    /// Reconstructs the true AS path for a 2-byte peer using RFC 6793 trailing-sequence
-    /// reconstruction. The last N ASNs in AS_PATH are replaced with the AS4_PATH values,
-    /// where N = min(AS_PATH length, AS4_PATH length). Internal for test coverage.
-    /// </summary>
-    internal static uint[] MergeAsPathWithAs4Path(uint[] asPath, uint[] as4Path)
-    {
-        if (as4Path.Length == 0)
-            return asPath;
-
-        if (as4Path.Length > asPath.Length)
-            throw new BgpNotificationException(BgpConstants.Error.UpdateMessageError, BgpConstants.SubError.Unspecific, "AS4_PATH longer than AS_PATH");
-
-        if (as4Path.Length == asPath.Length)
-            return as4Path;
-
-        var leadingCount = asPath.Length - as4Path.Length;
-        for (var i = 0; i < leadingCount; i++)
-        {
-            if (asPath[i] == BgpConstants.AsPath.AsTrans)
-                throw new BgpNotificationException(BgpConstants.Error.UpdateMessageError, BgpConstants.SubError.Unspecific, "Unresolved AS_TRANS in AS_PATH");
-        }
-
-        var merged = new uint[asPath.Length];
-        Array.Copy(asPath, 0, merged, 0, leadingCount);
-        Array.Copy(as4Path, 0, merged, leadingCount, as4Path.Length);
-
-        return merged;
     }
 
     /// <summary>
@@ -1129,75 +986,6 @@ public sealed class BgpSession : IDisposable
             Communities = communities,
             LargeCommunities = largeCommunities ?? []
         };
-
-    /// <summary>Sequence equality over a route's community array (set-equivalence within a batch).</summary>
-    private sealed class CommunitySetComparer : IEqualityComparer<uint[]>
-    {
-        public static readonly CommunitySetComparer Instance = new();
-
-        public bool Equals(uint[]? x, uint[]? y)
-        {
-            if (ReferenceEquals(x, y)) return true;
-            if (x is null || y is null || x.Length != y.Length) return false;
-            for (var i = 0; i < x.Length; i++)
-                if (x[i] != y[i]) return false;
-            return true;
-        }
-
-        public int GetHashCode(uint[] obj)
-        {
-            var hc = new HashCode();
-            foreach (var c in obj) hc.Add(c);
-            return hc.ToHashCode();
-        }
-    }
-
-    /// <summary>Sequence equality over a route's Large Community array (RFC 8092 triplets).</summary>
-    private sealed class LargeCommunitySetComparer : IEqualityComparer<(uint Global, uint Local1, uint Local2)[]>
-    {
-        public static readonly LargeCommunitySetComparer Instance = new();
-
-        public bool Equals((uint Global, uint Local1, uint Local2)[]? x, (uint Global, uint Local1, uint Local2)[]? y)
-        {
-            if (ReferenceEquals(x, y)) return true;
-            if (x is null || y is null || x.Length != y.Length) return false;
-            for (var i = 0; i < x.Length; i++)
-                if (x[i] != y[i]) return false;
-            return true;
-        }
-
-        public int GetHashCode((uint Global, uint Local1, uint Local2)[] obj)
-        {
-            var hc = new HashCode();
-            foreach (var c in obj) hc.Add(c);
-            return hc.ToHashCode();
-        }
-    }
-
-    /// <summary>
-    /// Composite sequence equality over a route's (regular, large) community pair, used to
-    /// partition a send batch that spans more than one community set.
-    /// </summary>
-    private sealed class CommunitySetPairComparer
-        : IEqualityComparer<(uint[] Communities, (uint Global, uint Local1, uint Local2)[] LargeCommunities)>
-    {
-        public static readonly CommunitySetPairComparer Instance = new();
-
-        public bool Equals(
-            (uint[] Communities, (uint Global, uint Local1, uint Local2)[] LargeCommunities) x,
-            (uint[] Communities, (uint Global, uint Local1, uint Local2)[] LargeCommunities) y) =>
-            CommunitySetComparer.Instance.Equals(x.Communities, y.Communities) &&
-            LargeCommunitySetComparer.Instance.Equals(x.LargeCommunities, y.LargeCommunities);
-
-        public int GetHashCode(
-            (uint[] Communities, (uint Global, uint Local1, uint Local2)[] LargeCommunities) obj)
-        {
-            var hc = new HashCode();
-            foreach (var c in obj.Communities) hc.Add(c);
-            foreach (var l in obj.LargeCommunities) hc.Add(l);
-            return hc.ToHashCode();
-        }
-    }
 
     private async Task SendUpdateBatchAsync(List<PathAttribute> attrs, List<IpPrefix> nlri)
     {
@@ -1517,7 +1305,7 @@ public sealed class BgpSession : IDisposable
         if (open.Version != BgpConstants.BgpVersion)
             throw new BgpNotificationException(BgpConstants.Error.OpenMessageError, BgpConstants.SubError.UnsupportedVersion, $"Unsupported BGP version: {open.Version}");
 
-        var malformedFourOctetAsnCapability = GetMalformedFourOctetAsnCapabilityData(open);
+        var malformedFourOctetAsnCapability = UpdateCodec.GetMalformedFourOctetAsnCapabilityData(open);
         if (malformedFourOctetAsnCapability.Length > 0)
         {
             throw new BgpNotificationException(
@@ -1556,27 +1344,6 @@ public sealed class BgpSession : IDisposable
             remoteRouteRefresh,
             holdTime,
             keepAliveInterval);
-    }
-
-    internal static byte[] GetMalformedFourOctetAsnCapabilityData(BgpOpenMessage open)
-    {
-        foreach (var cap in open.Capabilities)
-        {
-            if (cap.Code == BgpConstants.Capability.FourOctetAsn && cap.Data.Length != 4)
-                return [BgpConstants.Capability.FourOctetAsn, (byte)cap.Data.Length,
-                    ..cap.Data];
-        }
-
-        return [];
-    }
-
-    internal static void ValidateAggregatorReconstruction(uint? aggregatorAsn, uint? as4AggregatorAsn)
-    {
-        if (aggregatorAsn == BgpConstants.AsPath.AsTrans && as4AggregatorAsn is null)
-            throw new BgpNotificationException(BgpConstants.Error.UpdateMessageError, BgpConstants.SubError.Unspecific, "Missing AS4_AGGREGATOR for AGGREGATOR AS_TRANS");
-
-        if (!aggregatorAsn.HasValue && as4AggregatorAsn.HasValue)
-            throw new BgpNotificationException(BgpConstants.Error.UpdateMessageError, BgpConstants.SubError.Unspecific, "Missing AGGREGATOR attribute for AS4_AGGREGATOR");
     }
 
     #endregion
