+++ BGPLite.Tests/BgpSessionRfc6793Tests.cs
@@ -17,7 +17,7 @@ public class BgpSessionRfc6793Tests
     [Fact]
     public void MergeAsPathWithAs4Path_ReconstructsTrueSequence()
     {
-        var merged = BgpSession.MergeAsPathWithAs4Path([65010u, BgpConstants.AsPath.AsTrans, 65001u], [200000u, 65001u]);
+        var merged = UpdateCodec.MergeAsPathWithAs4Path([65010u, BgpConstants.AsPath.AsTrans, 65001u], [200000u, 65001u]);
 
         Assert.Equal([65010u, 200000u, 65001u], merged);
     }
@@ -26,7 +26,7 @@ public class BgpSessionRfc6793Tests
     public void MergeAsPathWithAs4Path_UnresolvedAsTrans_Throws()
     {
         var ex = Assert.Throws<BgpNotificationException>(() =>
-            BgpSession.MergeAsPathWithAs4Path([65010u, BgpConstants.AsPath.AsTrans, 65001u], [200000u]));
+            UpdateCodec.MergeAsPathWithAs4Path([65010u, BgpConstants.AsPath.AsTrans, 65001u], [200000u]));
 
         Assert.Equal(BgpConstants.Error.UpdateMessageError, ex.ErrorCode);
         Assert.Equal(BgpConstants.SubError.Unspecific, ex.SubErrorCode);
@@ -35,7 +35,7 @@ public class BgpSessionRfc6793Tests
     [Fact]
     public void MergeAsPathWithAs4Path_EmptyAs4Path_ReturnsAsPathUnchanged()
     {
-        var merged = BgpSession.MergeAsPathWithAs4Path([65010u, 65001u], []);
+        var merged = UpdateCodec.MergeAsPathWithAs4Path([65010u, 65001u], []);
 
         Assert.Equal([65010u, 65001u], merged);
     }
@@ -43,7 +43,7 @@ public class BgpSessionRfc6793Tests
     [Fact]
     public void MergeAsPathWithAs4Path_EqualLengths_ReturnsAs4Path()
     {
-        var merged = BgpSession.MergeAsPathWithAs4Path([BgpConstants.AsPath.AsTrans, 65001u], [200000u, 65001u]);
+        var merged = UpdateCodec.MergeAsPathWithAs4Path([BgpConstants.AsPath.AsTrans, 65001u], [200000u, 65001u]);
 
         Assert.Equal([200000u, 65001u], merged);
     }
@@ -52,7 +52,7 @@ public class BgpSessionRfc6793Tests
     public void MergeAsPathWithAs4Path_As4PathLongerThanAsPath_Throws()
     {
         var ex = Assert.Throws<BgpNotificationException>(() =>
-            BgpSession.MergeAsPathWithAs4Path([65001u], [200000u, 300000u]));
+            UpdateCodec.MergeAsPathWithAs4Path([65001u], [200000u, 300000u]));
 
         Assert.Equal(BgpConstants.Error.UpdateMessageError, ex.ErrorCode);
         Assert.Equal(BgpConstants.SubError.Unspecific, ex.SubErrorCode);
@@ -61,26 +61,26 @@ public class BgpSessionRfc6793Tests
     [Fact]
     public void ValidateAggregatorReconstruction_AsTransWithAs4Aggregator_DoesNotThrow()
     {
-        BgpSession.ValidateAggregatorReconstruction(BgpConstants.AsPath.AsTrans, 200000u);
+        UpdateCodec.ValidateAggregatorReconstruction(BgpConstants.AsPath.AsTrans, 200000u);
     }
 
     [Fact]
     public void ValidateAggregatorReconstruction_NonAsTransAggregator_DoesNotThrow()
     {
-        BgpSession.ValidateAggregatorReconstruction(65001u, null);
+        UpdateCodec.ValidateAggregatorReconstruction(65001u, null);
     }
 
     [Fact]
     public void ValidateAggregatorReconstruction_NullAggregator_DoesNotThrow()
     {
-        BgpSession.ValidateAggregatorReconstruction(null, null);
+        UpdateCodec.ValidateAggregatorReconstruction(null, null);
     }
 
     [Fact]
     public void ValidateAggregatorReconstruction_NullAggregatorWithAs4Aggregator_Throws()
     {
         var ex = Assert.Throws<BgpNotificationException>(() =>
-            BgpSession.ValidateAggregatorReconstruction(null, 200000u));
+            UpdateCodec.ValidateAggregatorReconstruction(null, 200000u));
 
         Assert.Equal(BgpConstants.Error.UpdateMessageError, ex.ErrorCode);
         Assert.Equal(BgpConstants.SubError.Unspecific, ex.SubErrorCode);
@@ -90,7 +90,7 @@ public class BgpSessionRfc6793Tests
     public void ValidateAggregatorReconstruction_AsTransWithoutAs4Aggregator_Throws()
     {
         var ex = Assert.Throws<BgpNotificationException>(() =>
-            BgpSession.ValidateAggregatorReconstruction(BgpConstants.AsPath.AsTrans, null));
+            UpdateCodec.ValidateAggregatorReconstruction(BgpConstants.AsPath.AsTrans, null));
 
         Assert.Equal(BgpConstants.Error.UpdateMessageError, ex.ErrorCode);
         Assert.Equal(BgpConstants.SubError.Unspecific, ex.SubErrorCode);
@@ -152,7 +152,7 @@ public class BgpSessionRfc6793Tests
     public void ValidateMandatoryAttributes_MissingRequiredAttribute_Throws()
     {
         var ex = Assert.Throws<BgpNotificationException>(() =>
-            BgpSession.ValidateMandatoryAttributes(false, true, true));
+            UpdateCodec.ValidateMandatoryAttributes(false, true, true));
 
         Assert.Equal(BgpConstants.Error.UpdateMessageError, ex.ErrorCode);
         Assert.Equal(BgpConstants.SubError.MissingWellKnownAttribute, ex.SubErrorCode);
@@ -162,7 +162,7 @@ public class BgpSessionRfc6793Tests
     public void ValidateMandatoryAttributes_MissingAsPath_Throws()
     {
         var ex = Assert.Throws<BgpNotificationException>(() =>
-            BgpSession.ValidateMandatoryAttributes(true, false, true));
+            UpdateCodec.ValidateMandatoryAttributes(true, false, true));
 
         Assert.Equal(BgpConstants.Error.UpdateMessageError, ex.ErrorCode);
         Assert.Equal(BgpConstants.SubError.MissingWellKnownAttribute, ex.SubErrorCode);
@@ -172,7 +172,7 @@ public class BgpSessionRfc6793Tests
     public void ValidateMandatoryAttributes_MissingNextHop_Throws()
     {
         var ex = Assert.Throws<BgpNotificationException>(() =>
-            BgpSession.ValidateMandatoryAttributes(true, true, false));
+            UpdateCodec.ValidateMandatoryAttributes(true, true, false));
 
         Assert.Equal(BgpConstants.Error.UpdateMessageError, ex.ErrorCode);
         Assert.Equal(BgpConstants.SubError.MissingWellKnownAttribute, ex.SubErrorCode);
@@ -186,7 +186,7 @@ public class BgpSessionRfc6793Tests
             Capabilities = [new BgpCapabilityInfo { Code = BgpConstants.Capability.FourOctetAsn, Data = [0x01, 0x02, 0x03] }]
         };
 
-        Assert.Equal([BgpConstants.Capability.FourOctetAsn, 3, 0x01, 0x02, 0x03], BgpSession.GetMalformedFourOctetAsnCapabilityData(open));
+        Assert.Equal([BgpConstants.Capability.FourOctetAsn, 3, 0x01, 0x02, 0x03], UpdateCodec.GetMalformedFourOctetAsnCapabilityData(open));
     }
 
     [Fact]
@@ -197,7 +197,7 @@ public class BgpSessionRfc6793Tests
             Capabilities = [new BgpCapabilityInfo { Code = BgpConstants.Capability.FourOctetAsn, Data = [] }]
         };
 
-        Assert.Equal([BgpConstants.Capability.FourOctetAsn, 0], BgpSession.GetMalformedFourOctetAsnCapabilityData(open));
+        Assert.Equal([BgpConstants.Capability.FourOctetAsn, 0], UpdateCodec.GetMalformedFourOctetAsnCapabilityData(open));
     }
 
     [Fact]
@@ -212,6 +212,6 @@ public class BgpSessionRfc6793Tests
             ]
         };
 
-        Assert.Equal([BgpConstants.Capability.FourOctetAsn, 2, 0x01, 0x02], BgpSession.GetMalformedFourOctetAsnCapabilityData(open));
+        Assert.Equal([BgpConstants.Capability.FourOctetAsn, 2, 0x01, 0x02], UpdateCodec.GetMalformedFourOctetAsnCapabilityData(open));
     }
 }
